refactoring of all progress bars into one parent component
