+++ src/exercises/exerciseTypes/match/Match.js
@@ -3,7 +3,7 @@ import * as s from "../Exercise.sc.js";
 import strings from "../../../i18n/definitions";
 import shuffle from "../../../assorted/fisherYatesShuffle";
 import { EXERCISE_TYPES } from "../../ExerciseTypeConstants.js";
-import LearningCycleIndicator from "../../LearningCycleIndicator.js";
+import BookmarkProgressBar from "../../progressBars/BookmarkProgressBar.js";
 import { SpeechContext } from "../../../contexts/SpeechContext.js";
 import NextNavigation from "../NextNavigation";
 import MatchInput from "./MatchInput.js";
@@ -204,7 +204,7 @@ export default function Match({
         {strings.matchWordWithTranslation}{" "}
       </div>
 
-      <LearningCycleIndicator
+      <BookmarkProgressBar
         bookmark={selectedBookmark ? selectedBookmark : bookmarksToStudy[0]}
         message={selectedBookmarkMessage}
         isHidden={selectedBookmark === undefined}
