+++ src/exercises/exerciseTypes/translateL2toL1/TranslateL2toL1.js
@@ -9,7 +9,7 @@ import { TranslatableText } from "../../../reader/TranslatableText.js";
 import { SpeechContext } from "../../../contexts/SpeechContext.js";
 import useSubSessionTimer from "../../../hooks/useSubSessionTimer.js";
 import BottomInput from "../BottomInput.js";
-import LearningCycleIndicator from "../../LearningCycleIndicator.js";
+import BookmarkProgressBar from "../../progressBars/BookmarkProgressBar.js";
 import { removePunctuation } from "../../../utils/text/preprocessing";
 
 // The user has to translate the L2 word in bold to their L1.
@@ -105,7 +105,7 @@ export default function TranslateL2toL1({
       <div className="headlineWithMoreSpace">
         {strings.translateL2toL1Headline}
       </div>
-      <LearningCycleIndicator
+      <BookmarkProgressBar
         bookmark={bookmarksToStudy[0]}
         message={messageToAPI}
       />
